requirements.txt
streamlit
pdfplumber
python-dotenv
google-generativeai
requests
streamlit-mermaid